refactor: 카페 휴무일 값타입 컬렉션에서 간접참조로 변경
