+++ src/main/java/com/birca/bircabackend/command/cafe/domain/DayOff.java
@@ -1,5 +1,6 @@
-package com.birca.bircabackend.command.cafe.domain.value;
+package com.birca.bircabackend.command.cafe.domain;
 
+import com.birca.bircabackend.common.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -7,19 +8,19 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
-@Embeddable
-@Table(uniqueConstraints = {@UniqueConstraint(
-        name = "uc_day_off_dayOffDate",
-        columnNames = "dayOffDate"
-)})
+@Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class DayOff {
+public class DayOff extends BaseEntity {
+
+    @Column(nullable = false)
+    private Long cafeId;
 
     @Column(nullable = false)
     private LocalDateTime dayOffDate;
 
-    public DayOff(LocalDateTime dayOffDate) {
+    public DayOff(Long cafeId, LocalDateTime dayOffDate) {
+        this.cafeId = cafeId;
         this.dayOffDate = dayOffDate;
     }
 }
