+++ src/main/java/com/birca/bircabackend/command/cafe/domain/DayOffRepository.java
@@ -1,0 +1,10 @@
+package com.birca.bircabackend.command.cafe.domain;
+
+import org.springframework.data.repository.Repository;
+
+public interface DayOffRepository extends Repository<DayOff, Long> {
+
+    void save(DayOff dayOff);
+
+    void deleteByCafeId(Long cafeId);
+}
